+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/B1ScanSpin.java
@@ -19,8 +19,8 @@ import org.openftc.easyopencv.OpenCvInternalCamera;
 import org.openftc.easyopencv.OpenCvPipeline;
 
 
-@Autonomous(name="Red scan & spin ")
-public class R1ScanSpinDelay extends LinearOpMode {
+@Autonomous
+public class B1ScanSpin extends LinearOpMode {
 
     private static final long SLEEP_10 = 10;
     private static final long SLEEP_25 = 25;
@@ -77,7 +77,7 @@ public class R1ScanSpinDelay extends LinearOpMode {
 
 
         waitForStart();
-        sleep(globalvar.delaytime);
+
 
         double FORWARD_SPEED = 0.5;
 
@@ -86,6 +86,7 @@ public class R1ScanSpinDelay extends LinearOpMode {
 
 
         while (opModeIsActive()) {
+
 
 
 
@@ -100,7 +101,7 @@ public class R1ScanSpinDelay extends LinearOpMode {
 
             spinAndComeBack();
             //stop();
-            robotMoveToShippingElement(6.5);
+            robotMoveToShippingElement(1.75);
 
 
             boolean positionShippingElement = scanShippingElement();
@@ -110,12 +111,12 @@ public class R1ScanSpinDelay extends LinearOpMode {
             if (positionShippingElement){
                 telemetry.addData("The position is",position);
                 telemetry.update();
-                moveToShippingHub(moveLength + 5, position);
+                moveToShippingHub(moveLength, position);
             }
 
             else {
 
-                moveToSecondDuck(9.5);
+                moveToSecondDuck(8);
                 positionShippingElement = scanShippingElement();
                 telemetry.addData("Is the shipping element present",positionShippingElement);
                 telemetry.update();
@@ -125,70 +126,40 @@ public class R1ScanSpinDelay extends LinearOpMode {
                     telemetry.addData("The position is",position);
                     telemetry.update();
 
-                    moveToShippingHub(moveLength - 4.5, position);
+                    moveToShippingHub(moveLength - 8, position);
                 }
                 else {
                     position = 3;
                     telemetry.addData("The position is",position);
                     telemetry.update();
 
-                    moveToShippingHub(moveLength - 1, position);
+                    moveToShippingHub(moveLength - 6, position);
                 }
             }
 
             placeFreightCorrectLocation(position);
             stop();
 
-
-
-            //switch - scan position one and two
-            //if not at one, then go to position two  if not at two, then it is at three
-            //if at one, then continue  if not at one, then go to position two
-            //if at position two, then continue
-
-            //Move forward/strafe toward shipping hub
-            //Place freight in correct location
-            //Finally go to warehouse
-
-            //sleep(5000);
-
-
-
-
         }
     }
 
     private void spinAndComeBack() {
 
-        //encoderDriveWithoutTime(.5,4,-4,4,-4);
-        //sleep(100);
-        encoderDriveWithoutTime(.5,-1,-1,-1,-1);
-        sleep(50);
-
-        encoderDriveWithoutTime(.5,-20,20,-20,20);
-        sleep(50);
-        encoderDriveWithoutTime(.1,-2.5,2.5,-2.5,2.5);
+        encoderDriveWithoutTime(.5,4,-4,4,-4);
         sleep(100);
-        //encoderDriveWithoutTime(.3,50,50,50,50);
-        //sleep(100);
-        spinRight(2.5,.4);
+        encoderDriveWithoutTime(.4,23,23,23,23);
         sleep(50);
-       //strafes left
-        encoderDriveWithoutTime(.5,11.5,-11.5,11.5,-11.5);
-        encoderDriveWithoutTime(.7,-6,-6,-6,-6);
-        turn90LeftMore();
-        turn90LeftMore();
+        encoderDriveWithoutTime(.1,3,3,3,3);
         sleep(50);
-        encoderDriveWithoutTime(0.3,-2.5,2.5,2.5,-2.5);
-        sleep(50);
-
-        /*encoderDriveWithoutTime(.5,-1,1,1,-1);
+        spinLeft(3,.4);
         sleep(100);
-        encoderDriveWithoutTime(.5,6,-6,6,-6);
+        encoderDriveWithoutTime(.5,-1,1,1,-1);
+        sleep(100);
+        encoderDriveWithoutTime(.5,5,-5,5,-5);
         sleep(100);
         encoderDriveWithoutTime(.5,-19,-19,-19,-19);
         sleep(100);
-        turn90Left(); */
+        turn90Left();
     }
 
     private void turn90Left() {
@@ -210,9 +181,9 @@ public class R1ScanSpinDelay extends LinearOpMode {
         //encoderDriveWithoutTime(0.5,4,4,4,4);
         //go forward
         //encoderDriveWithoutTime(0.5, 48,48,48,48);
-        encoderDriveWithoutTime(.8,8,8,8,8);
+        encoderDriveWithoutTime(.8,5,5,5,5);
         sleep(100);
-        turn90Left();
+        turn90Right();
         telemetry.addData("Before final move to warehouse","");
         telemetry.update();
         encoderDriveWithoutTime(.8,65,65,65,65);
@@ -223,20 +194,20 @@ public class R1ScanSpinDelay extends LinearOpMode {
 
         switch (position) {
             case 1:
-                //adjust arm position and drop cube
-                double len = 2;
-                encoderDriveWithoutTime(0.5, -10.5, -10.5, -10.5, -10.5);
-                encoderDriveWithoutTime(.4,-3,-3,-3,-3);
-
-
+                robot.Slider.setPower(1);
+                sleep(750);
+                robot.Slider.setPower(0.1);
+                sleep(100);
+                encoderDriveWithoutTime(0.5, -10.7, -10.7, -10.7, -10.7);
+                //encoderDriveWithoutTime(.4,-3,-3,-3,-3);
+                sleep(50);
                 robot.Intake1.setPower(0.8);
                 robot.Intake2.setPower(-0.8);
                 sleep(2000);
                 robot.Intake2.setPower(0);
                 robot.Intake1.setPower(0);
-
-
                 break;
+
             case 2:
                 robot.Slider.setPower(0.7);
                 sleep(450);
@@ -250,25 +221,24 @@ public class R1ScanSpinDelay extends LinearOpMode {
                 sleep(2000);
                 robot.Intake2.setPower(0);
                 robot.Intake1.setPower(0);
-
-                //adjust arm position and drop cube
                 break;
             default:
-                robot.Slider.setPower(1);
-                sleep(800);
-                robot.Slider.setPower(0.1);
-                sleep(100);
+                //adjust arm position and drop cube
+
+
+                //adjust arm position and drop cube
+                double len = 2;
                 encoderDriveWithoutTime(0.5, -10.5, -10.5, -10.5, -10.5);
                 encoderDriveWithoutTime(.4,-3,-3,-3,-3);
+
+
                 robot.Intake1.setPower(0.8);
                 robot.Intake2.setPower(-0.8);
                 sleep(2000);
                 robot.Intake2.setPower(0);
                 robot.Intake1.setPower(0);
-
-
-                //adjust arm position and drop c4ube
                 break;
+
 
         }
 
@@ -276,22 +246,21 @@ public class R1ScanSpinDelay extends LinearOpMode {
 
     private void moveToSecondDuck(double moveLength) {
 
-        encoderDriveWithoutTime(0.3, -moveLength, moveLength, -moveLength, moveLength);
+        encoderDriveWithoutTime(0.3, moveLength, -moveLength, moveLength, -moveLength);
         //encoderDriveWithTime(0.3,1,1,1,1,2);
     }
 
     private void moveToShippingHub(double moveLength, int position) {
 
-        encoderDriveWithoutTime(0.5, -moveLength - 2, moveLength + 2, -moveLength - 2, moveLength + 2);
+        encoderDriveWithoutTime(0.5, moveLength, -moveLength, moveLength, -moveLength);
         sleep(100);
         turn90LeftMore();
         telemetry.addData("Its working", "");
         telemetry.update();
         turn90LeftMore();
         sleep(100);
-        encoderDriveWithoutTime(0.5,4,-4,4,-4);
         //robot.Sweeper.setPosition(0);
-
+        //encoderDriveWithoutTime(0.5, -10.5, -10.5, -10.5, -10.5);
 
         //encoderDriveWithoutTime(.3, -6,6,-6,6);
         //placeFreightCorrectLocation(position);
@@ -302,40 +271,10 @@ public class R1ScanSpinDelay extends LinearOpMode {
         sleep(1000);
         int avg1 = pipeline.getAnalysis();
 
-        final int THRESHOLD = 155;//150;
+        final int THRESHOLD = 145;//150;
         telemetry.addData("Scan Threshhold:",avg1);
         telemetry.update();
         sleep(1000);
-
-
-        if(avg1 > THRESHOLD){
-            return true;
-        }else{
-            return false;
-        }
-
-
-    }
-    private boolean scanShippingElementTest() {
-
-        while (opModeIsActive()) {
-            sleep(100);
-            int avg1 = pipeline.getAnalysis();
-
-            final int THRESHOLD = 25;//150;
-            telemetry.addData("Scan Threshhold:",avg1);
-            telemetry.update();
-            sleep(2000);
-
-
-        }
-        sleep(100);
-        int avg1 = pipeline.getAnalysis();
-
-        final int THRESHOLD = 25;//150;
-        telemetry.addData("Scan Threshhold:",avg1);
-        telemetry.update();
-        sleep(5000);
 
 
         if(avg1 > THRESHOLD){
@@ -351,32 +290,9 @@ public class R1ScanSpinDelay extends LinearOpMode {
         telemetry.addData("Going toward Shipping element", tmpmoveLength);
         telemetry.update();
         encoderDriveWithoutTime(0.3, tmpmoveLength, tmpmoveLength, tmpmoveLength, tmpmoveLength);
-
         //encoderDriveWithTime(1,-1,-1,-1,-1,10);
     }
 
-
-
-    private void strafeLeft(double moveLength) {
-        telemetry.addData("Strafe Left: ", moveLength);
-        telemetry.update();
-        //encoderDriveWithoutTime(0.3, -moveLength, moveLength, -moveLength, moveLength);
-        encoderDriveWithTime(0.3,-1,1,-1,1,2);
-    }
-
-    private void strafeRight(double moveLength) {
-        telemetry.addData("Strafe Right: ", moveLength);
-        telemetry.update();
-        encoderDriveWithoutTime(0.3, moveLength, -moveLength, moveLength, -moveLength);
-    }
-
-    private void moveTowards(double moveLength) {
-
-        telemetry.addData("Move Length: ", moveLength);
-        telemetry.update();
-        encoderDriveWithoutTime(0.5, -moveLength, -moveLength, -moveLength, -moveLength);
-
-    }
 
     public void encoderDriveWithoutTime ( double speed,
                                           double Back_Left_Inches,
