+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/B2ScanPark.java
@@ -19,8 +19,8 @@ import org.openftc.easyopencv.OpenCvInternalCamera;
 import org.openftc.easyopencv.OpenCvPipeline;
 
 
-@Autonomous(name="Red scan & spin ")
-public class R1ScanSpinDelay extends LinearOpMode {
+@Autonomous
+public class B2ScanPark extends LinearOpMode {
 
     private static final long SLEEP_10 = 10;
     private static final long SLEEP_25 = 25;
@@ -77,7 +77,7 @@ public class R1ScanSpinDelay extends LinearOpMode {
 
 
         waitForStart();
-        sleep(globalvar.delaytime);
+
 
         double FORWARD_SPEED = 0.5;
 
@@ -91,16 +91,18 @@ public class R1ScanSpinDelay extends LinearOpMode {
 
 
 
+
+
+            double moveLength = 28;
+            int position = 1;
+
+            //spinAndComeBack();
+            //stop();
+            robotMoveToShippingElement(3);
             robot.Slider.setPower(1);
             sleep(400);
             robot.Slider.setPower(0);
             sleep(200);
-            double moveLength = 28;
-            int position = 1;
-
-            spinAndComeBack();
-            //stop();
-            robotMoveToShippingElement(6.5);
 
 
             boolean positionShippingElement = scanShippingElement();
@@ -110,85 +112,57 @@ public class R1ScanSpinDelay extends LinearOpMode {
             if (positionShippingElement){
                 telemetry.addData("The position is",position);
                 telemetry.update();
-                moveToShippingHub(moveLength + 5, position);
+                position = 2;
+                moveToShippingHub(moveLength - 1, position);
             }
 
             else {
 
-                moveToSecondDuck(9.5);
+                moveToSecondDuck(10);
                 positionShippingElement = scanShippingElement();
                 telemetry.addData("Is the shipping element present",positionShippingElement);
                 telemetry.update();
                 sleep(100);
                 if (positionShippingElement){
-                    position = 2;
+                    position = 1;
                     telemetry.addData("The position is",position);
                     telemetry.update();
 
-                    moveToShippingHub(moveLength - 4.5, position);
+                    moveToShippingHub(moveLength - 13, position);
                 }
                 else {
                     position = 3;
                     telemetry.addData("The position is",position);
                     telemetry.update();
 
-                    moveToShippingHub(moveLength - 1, position);
+                    moveToShippingHub(moveLength - 14, position);
                 }
             }
 
             placeFreightCorrectLocation(position);
+            moveToWarehouse();
             stop();
-
-
-
-            //switch - scan position one and two
-            //if not at one, then go to position two  if not at two, then it is at three
-            //if at one, then continue  if not at one, then go to position two
-            //if at position two, then continue
-
-            //Move forward/strafe toward shipping hub
-            //Place freight in correct location
-            //Finally go to warehouse
-
-            //sleep(5000);
-
-
-
 
         }
     }
 
     private void spinAndComeBack() {
 
-        //encoderDriveWithoutTime(.5,4,-4,4,-4);
-        //sleep(100);
-        encoderDriveWithoutTime(.5,-1,-1,-1,-1);
-        sleep(50);
-
-        encoderDriveWithoutTime(.5,-20,20,-20,20);
-        sleep(50);
-        encoderDriveWithoutTime(.1,-2.5,2.5,-2.5,2.5);
+        encoderDriveWithoutTime(.5,4,-4,4,-4);
         sleep(100);
-        //encoderDriveWithoutTime(.3,50,50,50,50);
-        //sleep(100);
-        spinRight(2.5,.4);
+        encoderDriveWithoutTime(.4,23,23,23,23);
         sleep(50);
-       //strafes left
-        encoderDriveWithoutTime(.5,11.5,-11.5,11.5,-11.5);
-        encoderDriveWithoutTime(.7,-6,-6,-6,-6);
-        turn90LeftMore();
-        turn90LeftMore();
+        encoderDriveWithoutTime(.1,3,3,3,3);
         sleep(50);
-        encoderDriveWithoutTime(0.3,-2.5,2.5,2.5,-2.5);
-        sleep(50);
-
-        /*encoderDriveWithoutTime(.5,-1,1,1,-1);
+        spinLeft(3,.4);
         sleep(100);
-        encoderDriveWithoutTime(.5,6,-6,6,-6);
+        encoderDriveWithoutTime(.5,-1,1,1,-1);
+        sleep(100);
+        encoderDriveWithoutTime(.5,5,-5,5,-5);
         sleep(100);
         encoderDriveWithoutTime(.5,-19,-19,-19,-19);
         sleep(100);
-        turn90Left(); */
+        turn90Left();
     }
 
     private void turn90Left() {
@@ -210,9 +184,9 @@ public class R1ScanSpinDelay extends LinearOpMode {
         //encoderDriveWithoutTime(0.5,4,4,4,4);
         //go forward
         //encoderDriveWithoutTime(0.5, 48,48,48,48);
-        encoderDriveWithoutTime(.8,8,8,8,8);
+        encoderDriveWithoutTime(.8,10,10,10,10);
         sleep(100);
-        turn90Left();
+        turn90Right();
         telemetry.addData("Before final move to warehouse","");
         telemetry.update();
         encoderDriveWithoutTime(.8,65,65,65,65);
@@ -223,26 +197,26 @@ public class R1ScanSpinDelay extends LinearOpMode {
 
         switch (position) {
             case 1:
-                //adjust arm position and drop cube
-                double len = 2;
-                encoderDriveWithoutTime(0.5, -10.5, -10.5, -10.5, -10.5);
-                encoderDriveWithoutTime(.4,-3,-3,-3,-3);
-
-
+                robot.Slider.setPower(1);
+                sleep(850);
+                robot.Slider.setPower(0.1);
+                sleep(100);
+                encoderDriveWithoutTime(0.5, -20.7, -20.7, -20.7, -20.7);
+                //encoderDriveWithoutTime(.4,-3,-3,-3,-3);
+                sleep(50);
                 robot.Intake1.setPower(0.8);
                 robot.Intake2.setPower(-0.8);
                 sleep(2000);
                 robot.Intake2.setPower(0);
                 robot.Intake1.setPower(0);
-
-
                 break;
+
             case 2:
                 robot.Slider.setPower(0.7);
-                sleep(450);
+                sleep(550);
                 robot.Slider.setPower(0.1);
                 sleep(100);
-                encoderDriveWithoutTime(0.5, -10.5, -10.5, -10.5, -10.5);
+                encoderDriveWithoutTime(0.5, -20, -20, -20, -20);
 
                 encoderDriveWithoutTime(.4,-3,-3,-3,-3);
                 robot.Intake1.setPower(0.8);
@@ -250,25 +224,24 @@ public class R1ScanSpinDelay extends LinearOpMode {
                 sleep(2000);
                 robot.Intake2.setPower(0);
                 robot.Intake1.setPower(0);
-
-                //adjust arm position and drop cube
                 break;
             default:
-                robot.Slider.setPower(1);
-                sleep(800);
-                robot.Slider.setPower(0.1);
-                sleep(100);
-                encoderDriveWithoutTime(0.5, -10.5, -10.5, -10.5, -10.5);
+                //adjust arm position and drop cube
+
+
+                //adjust arm position and drop cube
+                double len = 2;
+                encoderDriveWithoutTime(0.5, -16, -16, -16, -16);
                 encoderDriveWithoutTime(.4,-3,-3,-3,-3);
+
+
                 robot.Intake1.setPower(0.8);
                 robot.Intake2.setPower(-0.8);
                 sleep(2000);
                 robot.Intake2.setPower(0);
                 robot.Intake1.setPower(0);
-
-
-                //adjust arm position and drop c4ube
                 break;
+
 
         }
 
@@ -282,16 +255,15 @@ public class R1ScanSpinDelay extends LinearOpMode {
 
     private void moveToShippingHub(double moveLength, int position) {
 
-        encoderDriveWithoutTime(0.5, -moveLength - 2, moveLength + 2, -moveLength - 2, moveLength + 2);
+        encoderDriveWithoutTime(0.5, -moveLength, moveLength, -moveLength, moveLength);
         sleep(100);
         turn90LeftMore();
         telemetry.addData("Its working", "");
         telemetry.update();
         turn90LeftMore();
         sleep(100);
-        encoderDriveWithoutTime(0.5,4,-4,4,-4);
         //robot.Sweeper.setPosition(0);
-
+        //encoderDriveWithoutTime(0.5, -10.5, -10.5, -10.5, -10.5);
 
         //encoderDriveWithoutTime(.3, -6,6,-6,6);
         //placeFreightCorrectLocation(position);
@@ -302,7 +274,7 @@ public class R1ScanSpinDelay extends LinearOpMode {
         sleep(1000);
         int avg1 = pipeline.getAnalysis();
 
-        final int THRESHOLD = 155;//150;
+        final int THRESHOLD = 140;//150;
         telemetry.addData("Scan Threshhold:",avg1);
         telemetry.update();
         sleep(1000);
@@ -351,7 +323,6 @@ public class R1ScanSpinDelay extends LinearOpMode {
         telemetry.addData("Going toward Shipping element", tmpmoveLength);
         telemetry.update();
         encoderDriveWithoutTime(0.3, tmpmoveLength, tmpmoveLength, tmpmoveLength, tmpmoveLength);
-
         //encoderDriveWithTime(1,-1,-1,-1,-1,10);
     }
