+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/B2ScanParkDelay.java
@@ -19,8 +19,8 @@ import org.openftc.easyopencv.OpenCvInternalCamera;
 import org.openftc.easyopencv.OpenCvPipeline;
 
 
-@Autonomous(name = "Red Full 2")
-public class R2FullAuto extends LinearOpMode {
+@Autonomous
+public class B2ScanParkDelay extends LinearOpMode {
 
     private static final long SLEEP_10 = 10;
     private static final long SLEEP_25 = 25;
@@ -77,7 +77,7 @@ public class R2FullAuto extends LinearOpMode {
 
 
         waitForStart();
-
+        sleep(globalvar.delaytime);
 
         double FORWARD_SPEED = 0.5;
 
@@ -92,14 +92,15 @@ public class R2FullAuto extends LinearOpMode {
 
 
 
-            double moveLength = 35;
+
+            double moveLength = 28;
             int position = 1;
 
             //spinAndComeBack();
             //stop();
-            robotMoveToShippingElement(12);
+            robotMoveToShippingElement(3);
             robot.Slider.setPower(1);
-            sleep(600);
+            sleep(400);
             robot.Slider.setPower(0);
             sleep(200);
 
@@ -112,12 +113,12 @@ public class R2FullAuto extends LinearOpMode {
                 telemetry.addData("The position is",position);
                 telemetry.update();
                 position = 2;
-                moveToShippingHub(moveLength - 11.5, position);
+                moveToShippingHub(moveLength - 1, position);
             }
 
             else {
 
-                moveToSecondDuck(10.5);
+                moveToSecondDuck(10);
                 positionShippingElement = scanShippingElement();
                 telemetry.addData("Is the shipping element present",positionShippingElement);
                 telemetry.update();
@@ -127,14 +128,14 @@ public class R2FullAuto extends LinearOpMode {
                     telemetry.addData("The position is",position);
                     telemetry.update();
 
-                    moveToShippingHub(moveLength - 19.5, position);
+                    moveToShippingHub(moveLength - 13, position);
                 }
                 else {
                     position = 3;
                     telemetry.addData("The position is",position);
                     telemetry.update();
 
-                    moveToShippingHub(moveLength - 19.5, position);
+                    moveToShippingHub(moveLength - 14, position);
                 }
             }
 
@@ -142,56 +143,26 @@ public class R2FullAuto extends LinearOpMode {
             moveToWarehouse();
             stop();
 
-
-
-            //switch - scan position one and two
-            //if not at one, then go to position two  if not at two, then it is at three
-            //if at one, then continue  if not at one, then go to position two
-            //if at position two, then continue
-
-            //Move forward/strafe toward shipping hub
-            //Place freight in correct location
-            //Finally go to warehouse
-
-            //sleep(5000);
-
-
-
-
         }
     }
 
     private void spinAndComeBack() {
 
-        //encoderDriveWithoutTime(.5,4,-4,4,-4);
-        //sleep(100);
-        encoderDriveWithoutTime(.5,-1,-1,-1,-1);
-        sleep(50);
-
-        encoderDriveWithoutTime(.5,-20,20,-20,20);
-        sleep(50);
-        encoderDriveWithoutTime(.1,-2.5,2.5,-2.5,2.5);
+        encoderDriveWithoutTime(.5,4,-4,4,-4);
         sleep(100);
-        //encoderDriveWithoutTime(.3,50,50,50,50);
-        //sleep(100);
-        spinRight(2.5,.4);
+        encoderDriveWithoutTime(.4,23,23,23,23);
         sleep(50);
-       //strafes left
-        encoderDriveWithoutTime(.5,10.5,-10.5,10.5,-10.5);
-        encoderDriveWithoutTime(.7,-6,-6,-6,-6);
-        turn90LeftMore();
-        turn90LeftMore();
+        encoderDriveWithoutTime(.1,3,3,3,3);
         sleep(50);
-        encoderDriveWithoutTime(0.3,-2.5,2.5,2.5,-2.5);
-        sleep(50);
-
-        /*encoderDriveWithoutTime(.5,-1,1,1,-1);
+        spinLeft(3,.4);
         sleep(100);
-        encoderDriveWithoutTime(.5,6,-6,6,-6);
+        encoderDriveWithoutTime(.5,-1,1,1,-1);
+        sleep(100);
+        encoderDriveWithoutTime(.5,5,-5,5,-5);
         sleep(100);
         encoderDriveWithoutTime(.5,-19,-19,-19,-19);
         sleep(100);
-        turn90Left(); */
+        turn90Left();
     }
 
     private void turn90Left() {
@@ -213,10 +184,9 @@ public class R2FullAuto extends LinearOpMode {
         //encoderDriveWithoutTime(0.5,4,4,4,4);
         //go forward
         //encoderDriveWithoutTime(0.5, 48,48,48,48);
-        encoderDriveWithoutTime(.8,8,8,8,8);
+        encoderDriveWithoutTime(.8,10,10,10,10);
         sleep(100);
-        turn90Left();
-        encoderDriveWithoutTime(.4,-8,8,-8,8);
+        turn90Right();
         telemetry.addData("Before final move to warehouse","");
         telemetry.update();
         encoderDriveWithoutTime(.8,65,65,65,65);
@@ -227,26 +197,26 @@ public class R2FullAuto extends LinearOpMode {
 
         switch (position) {
             case 1:
-                //adjust arm position and drop cube
-                double len = 2;
-                encoderDriveWithoutTime(0.5, -8, -8, -8, -8);
-                encoderDriveWithoutTime(.4,-3,-3,-3,-3);
-
-
+                robot.Slider.setPower(1);
+                sleep(850);
+                robot.Slider.setPower(0.1);
+                sleep(100);
+                encoderDriveWithoutTime(0.5, -20.7, -20.7, -20.7, -20.7);
+                //encoderDriveWithoutTime(.4,-3,-3,-3,-3);
+                sleep(50);
                 robot.Intake1.setPower(0.8);
                 robot.Intake2.setPower(-0.8);
                 sleep(2000);
                 robot.Intake2.setPower(0);
                 robot.Intake1.setPower(0);
-
-
                 break;
+
             case 2:
                 robot.Slider.setPower(0.7);
                 sleep(550);
                 robot.Slider.setPower(0.1);
                 sleep(100);
-                encoderDriveWithoutTime(0.5, -10.5, -10.5, -10.5, -10.5);
+                encoderDriveWithoutTime(0.5, -20, -20, -20, -20);
 
                 encoderDriveWithoutTime(.4,-3,-3,-3,-3);
                 robot.Intake1.setPower(0.8);
@@ -254,25 +224,24 @@ public class R2FullAuto extends LinearOpMode {
                 sleep(2000);
                 robot.Intake2.setPower(0);
                 robot.Intake1.setPower(0);
-
-                //adjust arm position and drop cube
                 break;
             default:
-                robot.Slider.setPower(1);
-                sleep(750);
-                robot.Slider.setPower(0.1);
-                sleep(100);
-                encoderDriveWithoutTime(0.5, -8, -8, -8, -8);
+                //adjust arm position and drop cube
+
+
+                //adjust arm position and drop cube
+                double len = 2;
+                encoderDriveWithoutTime(0.5, -16, -16, -16, -16);
                 encoderDriveWithoutTime(.4,-3,-3,-3,-3);
+
+
                 robot.Intake1.setPower(0.8);
                 robot.Intake2.setPower(-0.8);
                 sleep(2000);
                 robot.Intake2.setPower(0);
                 robot.Intake1.setPower(0);
-
-
-                //adjust arm position and drop c4ube
                 break;
+
 
         }
 
@@ -280,22 +249,21 @@ public class R2FullAuto extends LinearOpMode {
 
     private void moveToSecondDuck(double moveLength) {
 
-        encoderDriveWithoutTime(0.3, moveLength, -moveLength, moveLength, -moveLength);
+        encoderDriveWithoutTime(0.3, -moveLength, moveLength, -moveLength, moveLength);
         //encoderDriveWithTime(0.3,1,1,1,1,2);
     }
 
     private void moveToShippingHub(double moveLength, int position) {
 
-        encoderDriveWithoutTime(0.5, moveLength + 2, -moveLength - 2, moveLength + 2, -moveLength - 2);
+        encoderDriveWithoutTime(0.5, -moveLength, moveLength, -moveLength, moveLength);
         sleep(100);
         turn90LeftMore();
         telemetry.addData("Its working", "");
         telemetry.update();
         turn90LeftMore();
         sleep(100);
-        encoderDriveWithoutTime(0.5,4,-4,4,-4);
         //robot.Sweeper.setPosition(0);
-
+        //encoderDriveWithoutTime(0.5, -10.5, -10.5, -10.5, -10.5);
 
         //encoderDriveWithoutTime(.3, -6,6,-6,6);
         //placeFreightCorrectLocation(position);
@@ -306,7 +274,7 @@ public class R2FullAuto extends LinearOpMode {
         sleep(1000);
         int avg1 = pipeline.getAnalysis();
 
-        final int THRESHOLD = 155;//150;
+        final int THRESHOLD = 140;//150;
         telemetry.addData("Scan Threshhold:",avg1);
         telemetry.update();
         sleep(1000);
@@ -355,7 +323,6 @@ public class R2FullAuto extends LinearOpMode {
         telemetry.addData("Going toward Shipping element", tmpmoveLength);
         telemetry.update();
         encoderDriveWithoutTime(0.3, tmpmoveLength, tmpmoveLength, tmpmoveLength, tmpmoveLength);
-
         //encoderDriveWithTime(1,-1,-1,-1,-1,10);
     }
