+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/R1SpinPark.java
@@ -21,7 +21,7 @@ import org.openftc.easyopencv.OpenCvPipeline;
 
 
 @Autonomous
-public class R1SpinAndPark extends LinearOpMode {
+public class R1SpinPark extends LinearOpMode {
 
     private static final long SLEEP_10 = 10;
     private static final long SLEEP_25 = 25;
