+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/globalvar.java
@@ -1,0 +1,5 @@
+package org.firstinspires.ftc.teamcode;
+
+public class globalvar {
+    static int delaytime=10000;
+}
